Renaming 'aux' directory to 'misc'
