+++ core/loaders/load_yaml.cpp
@@ -13,7 +13,7 @@
 //  ***********************************************************************************************
 
 #include "core/loaders/fastyamlloader.h"
-#include "core/aux/exception.h"
+#include "core/misc/exception.h"
 #include "qcr/base/debug.h"
 #include <functional>
