+++ core/pars/interpol_params.cpp
@@ -13,7 +13,7 @@
 //  ***********************************************************************************************
 
 #include "core/pars/interpol_params.h"
-#include "core/aux/settings.h"
+#include "core/misc/settings.h"
 #include "core/session.h"
 
 InterpolParams::InterpolParams()
