+++ core/typ/range.cpp
@@ -15,7 +15,7 @@
 //  ***********************************************************************************************
 
 #include "core/typ/range.h"
-#include "core/aux/exception.h"
+#include "core/misc/exception.h"
 #include "core/typ/json.h"
 #include "qcr/base/debug.h"
