+++ gui/dialogs/export_dfgram.cpp
@@ -13,8 +13,8 @@
 //  ***********************************************************************************************
 
 #include "gui/dialogs/export_dfgram.h"
-#include "core/aux/async.h"
-#include "core/aux/exception.h"
+#include "core/misc/async.h"
+#include "core/misc/exception.h"
 #include "core/session.h"
 #include "gui/dialogs/subdialog_file.h"
 #include "gui/dialogs/file_dialog.h"
